+++ FluidNC/esp32/uart.cpp
@@ -33,7 +33,7 @@ void uart_data_callback(uart_port_t uart_num, uint8_t* buf, uint32_t* len) {
     }
     *len = out;
 }
-void uart_register_input_pin(uint8_t uart_num, pinnum_t pinnum, InputPin* object) {
+void uart_register_input_pin(uint32_t uart_num, pinnum_t pinnum, InputPin* object) {
     objects[uart_num][pinnum] = object;
     last[uart_num]            = 0;
 }
@@ -47,7 +47,7 @@ static void uart_driver_n_install(void* arg) {
     }
 }
 
-void uart_init(uint8_t uart_num) {
+void uart_init(uint32_t uart_num) {
     // We init UARTs on core 0 so the interrupt handler runs there,
     // thus avoiding conflict with the StepTimer interrupt
     esp_ipc_call_blocking(0, uart_driver_n_install, &uart_num);
@@ -75,7 +75,7 @@ std::map<UartParity, uart_parity_t> parity_mode = {
     { UartParity::Odd, UART_PARITY_ODD },
 };
 
-void uart_mode(uint8_t uart_num, uint32_t baud, UartData dataBits, UartParity parity, UartStop stopBits) {
+void uart_mode(uint32_t uart_num, uint32_t baud, UartData dataBits, UartParity parity, UartStop stopBits) {
     uart_config_t conf;
     conf.source_clk          = UART_SCLK_APB;
     conf.baud_rate           = baud;
@@ -93,7 +93,7 @@ void uart_mode(uint8_t uart_num, uint32_t baud, UartData dataBits, UartParity pa
     }
 }
 
-bool uart_half_duplex(uint8_t uart_num) {
+bool uart_half_duplex(uint32_t uart_num) {
     uart_port_t port = (uart_port_t)uart_num;
     if (port) {
         return fnc_uart_set_mode(port, UART_MODE_RS485_HALF_DUPLEX) != ESP_OK;
@@ -102,7 +102,7 @@ bool uart_half_duplex(uint8_t uart_num) {
     }
 }
 
-int uart_read(uint8_t uart_num, uint8_t* buf, uint32_t len, uint32_t timeout_ms) {
+int uart_read(uint32_t uart_num, uint8_t* buf, uint32_t len, uint32_t timeout_ms) {
     uart_port_t port = (uart_port_t)uart_num;
     if (port) {
         return fnc_uart_read_bytes(port, buf, len, timeout_ms);
@@ -110,7 +110,7 @@ int uart_read(uint8_t uart_num, uint8_t* buf, uint32_t len, uint32_t timeout_ms)
         return uart_read_bytes(port, buf, len, timeout_ms);
     }
 }
-int uart_write(uint8_t uart_num, const uint8_t* buf, size_t len) {
+int uart_write(uint32_t uart_num, const uint8_t* buf, size_t len) {
     uart_port_t port = (uart_port_t)uart_num;
     if (port) {
         return fnc_uart_write_bytes(port, buf, len);
@@ -118,15 +118,15 @@ int uart_write(uint8_t uart_num, const uint8_t* buf, size_t len) {
         return uart_write_bytes(port, buf, len);
     }
 }
-void uart_xon(uint8_t uart_num) {
+void uart_xon(uint32_t uart_num) {
     uart_port_t port = (uart_port_t)uart_num;
     uart_ll_force_xon(port);
 }
-void uart_xoff(uint8_t uart_num) {
+void uart_xoff(uint32_t uart_num) {
     uart_port_t port = (uart_port_t)uart_num;
     uart_ll_force_xoff(port);
 }
-void uart_sw_flow_control(uint8_t uart_num, bool on, uint32_t xon_threshold, uint32_t xoff_threshold) {
+void uart_sw_flow_control(uint32_t uart_num, bool on, uint32_t xon_threshold, uint32_t xoff_threshold) {
     if (xon_threshold <= 0) {
         xon_threshold = 126;
     }
@@ -140,7 +140,7 @@ void uart_sw_flow_control(uint8_t uart_num, bool on, uint32_t xon_threshold, uin
         uart_set_sw_flow_ctrl(port, on, xon_threshold, xoff_threshold);
     }
 }
-bool uart_pins(uint8_t uart_num, pinnum_t tx_pin, pinnum_t rx_pin, pinnum_t rts_pin, pinnum_t cts_pin) {
+bool uart_pins(uint32_t uart_num, pinnum_t tx_pin, pinnum_t rx_pin, pinnum_t rts_pin, pinnum_t cts_pin) {
     uart_port_t port = (uart_port_t)uart_num;
     if (port) {
         return fnc_uart_set_pin(uart_num, tx_pin, rx_pin, rts_pin, cts_pin) != ESP_OK;
@@ -148,14 +148,14 @@ bool uart_pins(uint8_t uart_num, pinnum_t tx_pin, pinnum_t rx_pin, pinnum_t rts_
         return uart_set_pin(uart_num, tx_pin, rx_pin, rts_pin, cts_pin) != ESP_OK;
     }
 }
-int uart_bufavail(uint8_t uart_num) {
+int uart_bufavail(uint32_t uart_num) {
 #ifdef UART_HW_FIFO_LEN  // Apparently not all uarts have the same FIFO buffer.
     return UART_HW_FIFO_LEN(uart_num) - uart_buflen(uart_num);
 #else
     return UART_FIFO_LEN - uart_buflen(uart_num);
 #endif
 }
-int uart_buflen(uint8_t uart_num) {
+int uart_buflen(uint32_t uart_num) {
     size_t      size;
     uart_port_t port = (uart_port_t)uart_num;
     if (port) {
@@ -165,7 +165,7 @@ int uart_buflen(uint8_t uart_num) {
     }
     return size;
 }
-void uart_discard_input(uint8_t uart_num) {
+void uart_discard_input(uint32_t uart_num) {
     uart_port_t port = (uart_port_t)uart_num;
     if (port) {
         fnc_uart_flush_input(port);
@@ -173,7 +173,7 @@ void uart_discard_input(uint8_t uart_num) {
         uart_flush_input(port);
     }
 }
-bool uart_wait_output(uint8_t uart_num, uint32_t timeout_ms) {
+bool uart_wait_output(uint32_t uart_num, uint32_t timeout_ms) {
     uart_port_t port = (uart_port_t)uart_num;
     if (port) {
         return fnc_uart_wait_tx_done(port, timeout_ms) != ESP_OK;
