+++ FluidNC/src/UartChannel.h
@@ -12,8 +12,8 @@ private:
     Lineedit* _lineedit;
     Uart*     _uart;
 
-    int32_t _uart_num           = 0;
-    int32_t _report_interval_ms = 0;
+    uint32_t _uart_num           = 0;
+    int32_t  _report_interval_ms = 0;
 
     static constexpr int _ack_timeout = 2000;
 
@@ -33,14 +33,14 @@ public:
     int read() override;
 
     // Channel methods
-    int     rx_buffer_available() override;
-    void    flushRx() override;
-    size_t  timedReadBytes(char* buffer, size_t length, TickType_t timeout);
-    size_t  timedReadBytes(uint8_t* buffer, size_t length, TickType_t timeout) { return timedReadBytes((char*)buffer, length, timeout); };
-    bool    realtimeOkay(char c) override;
-    bool    lineComplete(char* line, char c) override;
-    uint8_t uart_num() { return _uart_num; }
-    Uart*   uart() { return _uart; }
+    int      rx_buffer_available() override;
+    void     flushRx() override;
+    size_t   timedReadBytes(char* buffer, size_t length, TickType_t timeout);
+    size_t   timedReadBytes(uint8_t* buffer, size_t length, TickType_t timeout) { return timedReadBytes((char*)buffer, length, timeout); };
+    bool     realtimeOkay(char c) override;
+    bool     lineComplete(char* line, char c) override;
+    uint32_t uart_num() { return _uart_num; }
+    Uart*    uart() { return _uart; }
 
     bool setAttr(uint8_t index, bool* valuep, const std::string& s);
